+++ source/scan-tool-mso/ChangeLog.md
@@ -1,4 +1,8 @@
 # ChangeLog for scan-tool-mso
 
+## Version 0.02 (2016-02-20)
+  - add option to set custom certificate file for peer verification, similar
+    to curl's --cacert option
+
 ## Version 0.01 (2016-02-02)
   - initial version (reduced and adjusted copy of scan-tool for VirusTotal)